refactor: super duper refactor, cleaner and much beautiful code

+++ Core/Files/Services/DeleteFileService.cs
@@ -1,6 +1,5 @@
 using Core.Providers;
-using Core.Replications.Events;
-using Core.Storages;
+using Core.Providers.Events;
 using DotNetCore.CAP;
 
 namespace Core.Files.Services;
@@ -9,30 +8,38 @@ internal sealed class DeleteFileService : IDeleteFileService
 {
     private readonly IFileRepository _fileRepository;
     private readonly IProviderRepository _providerRepository;
-    private readonly IStorageService _storageService;
+    private readonly IStorageServiceLocator _storageServiceLocator;
     private readonly ICapPublisher _capPublisher;
 
-    public DeleteFileService(IFileRepository fileRepository, IProviderRepository providerRepository, IStorageService storageService, ICapPublisher capPublisher)
+    public DeleteFileService(IFileRepository fileRepository, IProviderRepository providerRepository, ICapPublisher capPublisher, IStorageServiceLocator storageServiceLocator)
     {
         _fileRepository = fileRepository;
         _providerRepository = providerRepository;
-        _storageService = storageService;
         _capPublisher = capPublisher;
+        _storageServiceLocator = storageServiceLocator;
     }
 
     public async Task DeleteAsync(string link, CancellationToken cancellationToken = default)
     {
         var fileId = IdLink.Parse(link);
         var file = await _fileRepository.FindAsync(fileId, cancellationToken);
-        await _storageService.DeleteAsync(file!.Path, file.Name);
-        file.Locations.Remove(file.Locations.First(location => location.Provider == _storageService.Provider));
+        if (file is null)
+        {
+            throw new FileNotFoundException();
+        }
+
+        var storageService = await _storageServiceLocator.LocatePrimaryAsync(cancellationToken);
+        await storageService.DeleteAsync(file.Path, file.Name);
+        file.Locations.Remove(file.Locations.First(location => location.Provider == storageService.Name));
         if (file.Locations.Count == 0)
         {
             await _fileRepository.DeleteAsync(file, cancellationToken);
-            return;
+        }
+        else
+        {
+            await _fileRepository.UpdateAsync(file, cancellationToken);
         }
 
-        await _fileRepository.UpdateAsync(file, cancellationToken);
         foreach (var provider in await _providerRepository.FindAsync(cancellationToken))
         {
             await Task.Run(() =>
